pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agripipe"
version = "0.1.0"
description = "Data pipeline per dati agronomici: Excel → tensor PyTorch"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["agritech", "data-pipeline", "pytorch", "machine-learning"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pandas>=2.1",
    "numpy>=1.26",
    "openpyxl>=3.1",
    "torch>=2.1",
    "scikit-learn>=1.3",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "typer>=0.9",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=24.0",
    "ruff>=0.3",
    "pre-commit>=3.6",
]

[project.scripts]
agripipe = "agripipe.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=agripipe --cov-report=term-missing"
